+++ src/app/CliTools/Console/Command/Docker/CliCommand.php
@@ -20,7 +20,7 @@ namespace CliTools\Console\Command\Docker;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-use CliTools\Console\Builder\RemoteCommandBuilder;
+use CliTools\Console\Shell\CommandBuilder\RemoteCommandBuilder;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;