+++ src/app/CliTools/Console/Command/Mysql/BackupCommand.php
@@ -21,8 +21,9 @@ namespace CliTools\Console\Command\Mysql;
  */
 
 use CliTools\Database\DatabaseConnection;
-use CliTools\Console\Builder\CommandBuilder;
-use CliTools\Console\Builder\CommandBuilderInterface;
+use CliTools\Console\Shell\CommandBuilder\CommandBuilder;
+use CliTools\Console\Shell\CommandBuilder\CommandBuilderInterface;
+use CliTools\Utility\FilterUtility;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
@@ -142,18 +143,11 @@ class BackupCommand extends \CliTools\Console\Command\AbstractCommand {
             throw new \RuntimeException('MySQL dump filters "' . $filter . '" not available"');
         }
 
+        $this->output->writeln('<comment>Using filter "' . $filter . '"</comment>');
+
         // Get filtered tables
         $tableList = DatabaseConnection::tableList($database);
-
-        $tableListFiltered = array();
-        foreach ($tableList as $table) {
-            foreach ($filterList as $filter) {
-                if (preg_match($filter, $table)) {
-                    continue 2;
-                }
-            }
-            $tableListFiltered[] = $table;
-        }
+        $tableList = FilterUtility::mysqlTableFilter($tableList, $filterList);
 
         // Dump only structure
         $commandStructure = clone $command;
@@ -163,10 +157,10 @@ class BackupCommand extends \CliTools\Console\Command\AbstractCommand {
         $commandData = clone $command;
         $commandData
             ->addArgument('--no-create-info')
-            ->addArgumentList($tableListFiltered);
+            ->addArgumentList($tableList);
 
         // Combine both commands to one
-        $command = new \CliTools\Console\Builder\OutputCombineCommandBuilder();
+        $command = new \CliTools\Console\Shell\CommandBuilder\OutputCombineCommandBuilder();
         $command
             ->addCommandForCombinedOutput($commandStructure)
             ->addCommandForCombinedOutput($commandData);