+++ src/app/CliTools/Console/Command/Mysql/RestoreCommand.php
@@ -24,7 +24,7 @@ use CliTools\Database\DatabaseConnection;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use CliTools\Console\Builder\CommandBuilder;
+use CliTools\Console\Shell\CommandBuilder\CommandBuilder;
 
 class RestoreCommand extends \CliTools\Console\Command\AbstractCommand {