+++ src/app/CliTools/Console/Command/System/StartupCommand.php
@@ -21,8 +21,8 @@ namespace CliTools\Console\Command\System;
  */
 
 use CliTools\Database\DatabaseConnection;
-use CliTools\Console\Builder\CommandBuilder;
-use CliTools\Console\Builder\SelfCommandBuilder;
+use CliTools\Console\Shell\CommandBuilder\CommandBuilder;
+use CliTools\Console\Shell\CommandBuilder\SelfCommandBuilder;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;