+++ src/app/CliTools/Console/Shell/CommandBuilder/SelfCommandBuilder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CliTools\Console\Builder;
+namespace CliTools\Console\Shell\CommandBuilder;
 
 /*
  * CliTools Command
@@ -20,8 +20,7 @@ namespace CliTools\Console\Builder;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-class FullSelfCommandBuilder extends CommandBuilder {
-
+class SelfCommandBuilder extends CommandBuilder {
 
     /**
      * Initalized command
@@ -31,13 +30,15 @@ class FullSelfCommandBuilder extends CommandBuilder {
 
         $arguments = $_SERVER['argv'];
 
+        // Check if command is run inside PHAR
         if (\Phar::running()) {
-            // running as phar
+            // PHAR version
             $this->setCommand(array_shift($arguments));
         } elseif (!empty($_SERVER['_'])) {
+            // Plain PHP version
             if ($_SERVER['argv'][0] !== $_SERVER['_']) {
                 $this->setCommand($_SERVER['_']);
-                $this->addArgument(array_shift($arguments));
+                $this->addArgument(reset($arguments));
             }
         }
 
@@ -46,8 +47,5 @@ class FullSelfCommandBuilder extends CommandBuilder {
             $this->setCommand('php');
             $this->addArgument($_SERVER['PHP_SELF']);
         }
-
-        // All other arguments
-        $this->addArgumentList($arguments);
     }
 }