+++ src/app/CliTools/Exception/CommandExecutionException.php
@@ -20,8 +20,8 @@ namespace CliTools\Exception;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-use CliTools\Console\Builder\CommandBuilder;
-use CliTools\Console\Builder\CommandBuilderInterface;
+use CliTools\Console\Shell\CommandBuilder\CommandBuilder;
+use CliTools\Console\Shell\CommandBuilder\CommandBuilderInterface;
 
 class CommandExecutionException extends \RuntimeException {