+++ src/app/CliTools/Utility/UnixUtility.php
@@ -20,7 +20,7 @@ namespace CliTools\Utility;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-use CliTools\Console\Builder\CommandBuilder;
+use CliTools\Console\Shell\CommandBuilder\CommandBuilder;
 
 abstract class UnixUtility {