+++ CHANGELOG.md
@@ -7,6 +7,8 @@ CliTools Changelog
 - Added `make` (auto search for Makefile in tree)
 - Added `php:composer` (auto search for composer.yml in tree)
 - Added `mysql:convert` for automatic changing charset and collation of one database
+- Added `sync:backup` for backup to a shared server (reads clisync.yml)
+- Added `sync:restore` for restore from a shared server (reads clisync.yml)
 - Fixed some issues
 
 1.9.0 - 2015-05-06